+++ src/html/Form.php
@@ -24,7 +24,7 @@ class Form extends Tag
         return $this->attribute('method', strtolower($method));
     }
 
-    #[\Override]
+    #[Override]
     protected function before(): string
     {
         $this->attributes['method'] ??= 'post';