+++ src/widgets/grids/columns/RelativeTimeColumn.php
@@ -6,19 +6,20 @@ namespace davidhirtz\yii2\skeleton\widgets\grids\columns;
 
 use Closure;
 use davidhirtz\yii2\skeleton\helpers\ArrayHelper;
-use davidhirtz\yii2\timeago\Timeago;
+use davidhirtz\yii2\skeleton\html\custom\RelativeTime;
 use Override;
+use Stringable;
 use yii\base\Model;
 
-class TimeagoColumn extends LinkColumn
+class RelativeTimeColumn extends LinkColumn
 {
     public array|null|Closure $contentAttributes = ['class' => 'text-nowrap'];
     protected string $format = 'raw';
 
     #[Override]
-    protected function getValue(array|Model $model, string|int $key, int $index): ?string
+    protected function getValue(array|Model $model, string|int $key, int $index): Stringable
     {
         $timestamp = ArrayHelper::getValue($model, $this->property);
-        return $timestamp ? Timeago::tag($timestamp) : null;
+        return RelativeTime::make()->value($timestamp);
     }
 }